add qa

+++ src/main/java/com/example/tmnewa/entity/qa/QADesignItem.java
@@ -22,7 +22,7 @@ public class QADesignItem {
     private String name;
 
     @Column(name = "score")
-    private Float score;
+    private int score;
 
 
     private Integer seq;
